requirements.txt
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.1
coverage==4.2
coveralls==1.1
docopt==0.6.2
funcsigs==1.0.2
mock==2.0.0
nose==1.3.7
pbr==1.10.0
python-dateutil==2.6.0
requests==2.12.4
simplejson==3.10.0
six==1.10.0
slackclient==1.0.4
sophiabus230==0.2
websocket-client==0.40.0
wit==4.2.0
